Reorganize file structure.
