+++ src/pages/Home/components/ContentRoutes/components/ProtectedRoute/ProtectedRoute.js
@@ -2,7 +2,8 @@
 
 import { useContext } from "react";
 import { Navigate } from "react-router-dom";
-import { UserContext } from "../../App";
+import { UserContext } from "../../../../../../containers/CognitoContextProvider/CognitoContextProvider";
+
 
 const ProtectedRoute = (props) => {
   const { user } = useContext(UserContext);
