+++ src/pages/Home/components/Editor/components/Controls.js
