+++ src/pages/Home/components/ProfileContent/ProfileContent.js
@@ -1,12 +1,12 @@
 // Services
-import * as profileServices from "../../services/profileServices";
-import * as userUtils from "../../utils/userUtils";
+import * as profileServices from "../../../../services/profileServices";
+import * as userUtils from "../../../../utils/userUtils";
 
 // Components / hooks
 import { useEffect, useState, useContext } from "react";
-import Card from "../Card/Card";
+import Card from "../../../../components/Card/Card";
 import { useParams } from "react-router-dom";
-import { UserContext } from "../../App";
+import { UserContext } from "../../../../containers/CognitoContextProvider/CognitoContextProvider";
 import { Grid } from "@mui/material";
 
 const ProfileContent = () => {
